refactor: move persitence factory to separate file
