+++ src/agent/agent_queue/include/persistence_factory.hpp
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <persistence.h>
+
+#include <any>
+#include <memory>
+#include <string>
+#include <vector>
+
+class PersistenceFactory
+{
+public:
+    static std::unique_ptr<Persistence> createPersistence(const std::string& type, const std::vector<std::any>& args);
+};
